extracted stechzeiten out of the databasse module
